+++ clients/c2c/src/rest-api/types/index.ts
@@ -1,2 +1,1 @@
 export * from './get-c2-ctrade-history-response';
-export * from './get-c2-ctrade-history-response-data-inner';